Add details inside the SOA page
The detail include the title, subject, etc. It used the tabularx
environment because we don't want the long title to exceed the page
(i.e. we want the table to automatically to have new line if the text is
too long). I have also committed the changes inside the markdown file,
which mainly is just the extension of the previous commit to include new
variables that are used inside the template.